backend/requirements.txt
arrow==1.3.0
asgiref==3.9.2
binaryornot==0.4.4
boto3==1.40.76
botocore==1.40.76
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.1
cookiecutter==2.6.0
cryptography==44.0.3
dj-database-url==3.0.1
Django==5.2.7
django-browser-reload==1.18.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
django-tailwind==4.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
firebase-admin==6.8.0
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
honcho==2.0.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.5
requests==2.32.3
rich==14.0.0
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.11.0
